Rename type for display. Add error correction level to type.
