+++ TODO.md
@@ -1,8 +1,6 @@
 # TODO
 
 * Remove dependency on Compat (isnothing)
-* Rename package (QRCodes.jl or perhaps QuickResponse.jl)
-* Rename/(add multiple?) types for display (QR or QRCode, parametric for level?)
 * Fix commented out functions (export to png)
 * Fix commented out tests
 * Add type display related tests (how?)
