+++ src/types.jl
@@ -1,16 +1,18 @@
-struct QRC # Avoid name clash with qr factorization and QRCode
+struct QRCode{T} # Avoid name clash with qr factorization and QRCode
     s::String
+    eclevel::T
 end
+QRCode(s) = QRCode(s, Medium())
 
-function Base.show(io::IO, ::MIME"text/plain", s::QRC)
-    qrm = plainqr(s.s)
+function Base.show(io::IO, ::MIME"text/plain", s::QRCode)
+    qrm = qrcode(s.s, s.eclevel)
     print(io, UnicodePlots.heatmap(qrm;
                         padding=0,margin=0,border=:none,
                         width=size(qrm,2),colormap=[(1,1,1),(0,0,0)],labels=false))
 end
 
-function Base.show(io::IO, ::MIME"image/png", s::QRC)
-    qrm = plainqr(s.s)
+function Base.show(io::IO, ::MIME"image/png", s::QRCode)
+    qrm = qrcode(s.s, s.eclevel)
     scale = Int(round(200/size(qrm,1)))
     qrm = Gray.(repeat(.!qrm, inner=(scale,scale)))
     png = IOBuffer()
